Refactor: 뜨거운 뒹글러 목록 조회 User 디렉토리로 이동
- 기존 dooingle에서 조회 기능 관련 클래스 삭제
- StopWatch 어노테이션 제거
