--- src/main/kotlin/com/dooingle/domain/dooingle/repository/DooingleCountQueryDslRepository.kt
@@ -1,7 +1,0 @@
-package com.dooingle.domain.dooingle.repository
-
-import com.dooingle.domain.dooingle.dto.DooinglerResponse
-
-interface DooingleCountQueryDslRepository {
-    fun getHighCountDooinglers(): List<DooinglerResponse>
-}
