--- src/main/kotlin/com/dooingle/domain/dooingle/repository/DooingleQueryDslRepository.kt
@@ -1,7 +1,0 @@
-package com.dooingle.domain.dooingle.repository
-
-import com.dooingle.domain.dooingle.dto.DooinglerResponse
-
-interface DooingleQueryDslRepository {
-    fun getHotDooinglerList(): List<DooinglerResponse>
-}
