--- src/main/kotlin/com/dooingle/domain/dooingle/repository/DooingleRepository.kt
@@ -1,6 +1,0 @@
-package com.dooingle.domain.dooingle.repository
-
-import com.dooingle.domain.dooingle.model.Dooingle
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface DooingleRepository : JpaRepository<Dooingle, Long>, DooingleQueryDslRepository
