+++ src/main/kotlin/com/dooingle/domain/user/controller/UserController.kt
@@ -1,7 +1,7 @@
-package com.dooingle.domain.dooingle.controller
+package com.dooingle.domain.user.controller
 
-import com.dooingle.domain.dooingle.dto.DooinglerResponse
-import com.dooingle.domain.dooingle.service.DooingleService
+import com.dooingle.domain.user.dto.DooinglerResponse
+import com.dooingle.domain.user.service.UserService
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
@@ -9,13 +9,12 @@ import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
-@RequestMapping("/api")
-class DooingleController(
-    private val dooingleService: DooingleService
+@RequestMapping("/api/users")
+class UserController(
+    private val userService: UserService
 ) {
-    @GetMapping("/users")
+    @GetMapping
     fun GetDooinglerList(@RequestParam(required = false) condition: String?): ResponseEntity<List<DooinglerResponse>> {
-        return ResponseEntity.ok().body(dooingleService.getDooinglerList(condition))
+        return ResponseEntity.ok().body(userService.getDooinglerList(condition))
     }
-
 }
