+++ src/main/kotlin/com/dooingle/domain/user/dto/DooinglerResponse.kt
@@ -1,4 +1,4 @@
-package com.dooingle.domain.dooingle.dto
+package com.dooingle.domain.user.dto
 
 data class DooinglerResponse(
     val userId: Long,
