+++ src/main/kotlin/com/dooingle/domain/user/repository/DooingleCountQueryDslRepository.kt
@@ -1,0 +1,7 @@
+package com.dooingle.domain.user.repository
+
+import com.dooingle.domain.user.dto.DooinglerResponse
+
+interface DooingleCountQueryDslRepository {
+    fun getHighCountDooinglers(): List<DooinglerResponse>
+}
